HP-556 Extract parts of audit logging into its own file
This enables importing audit logging system from multiple places, as the
signal connections aren't among them anymore.
